Store Station Database update timestimes in the databsae itself
Previously it was stored in Settings.Default - but that is application
specific and separated from the database (so if you delete the database
the timestamps are still there).
So we store the timestamps of the web JSON source in the database itself.

There is an additional timestamp for the database itself - that is
probably not needed and should be removed.